requirements/base.txt
Django==4.1.3
numpy==1.24.3
python-dotenv==1.0.0
psycopg2==2.9.9
web3==6.15.1
djangorestframework==3.15.1
django-cors-headers==4.3.1 
pillow==10.3.0
